Principio SOLID
Se agrego el principio SOLID para el CAI.
