+++ codelabBackEnd/src/services/cai/caiErrors.js
@@ -1,0 +1,5 @@
+export function buildCaiError(message, status = 400) {
+  const err = new Error(message);
+  err.status = status;
+  return err;
+}
